requirements.txt
motion-python
streamlit
openai
instructor
rich
python-dotenv